// web-application/src/types-openapi/api/models/ListBookingCourtOccurrenceResponse.ts
export interface ListBookingCourtOccurrenceResponse {
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    bookingCourtId?: string;
    /**
     * 
     * @type {Date}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    date?: Date;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    startTime?: string;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    endTime?: string;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    note?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    customerName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    courtName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    courtId?: string;
    /**
     * 
     * @type {number}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    customerId?: number;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    startDate?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBookingCourtOccurrenceResponse
     */
    endDate?: string | null;
}